Progress on inter-solver diffusion calculations.
git-svn-id: https://svn.code.sf.net/p/moose/code/moose/branches/buildQ@4348 d1858a74-f715-0410-99c9-c66fe9cce3d6

Former-commit-id: 36fa414333284fa02a92fddd843060ed2fc7ef88

+++ mesh/CubeMesh.h
@@ -308,7 +308,7 @@ class CubeMesh: public ChemMesh
 		 * diffusion rates across junctions.
 		 * Indexed from zero.
 		 */
-		vector< unsigned int > extendedMeshEntrySize_;
+		vector< double > extendedMeshEntrySize_;
 };
 
 #endif	// _CUBE_MESH_H
